+++ src/BigInteger.php
@@ -259,8 +259,8 @@ final readonly class BigInteger extends BigNumber
      *
      * Using the default random bytes generator, this method is suitable for cryptographic use.
      *
-     * @param BigNumber|int|float|string   $min                  The lower bound. Must be convertible to a BigInteger.
-     * @param BigNumber|int|float|string   $max                  The upper bound. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string         $min                  The lower bound. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string         $max                  The upper bound. Must be convertible to a BigInteger.
      * @param (callable(int): string)|null $randomBytesGenerator A function that accepts a number of bytes, and returns
      *                                                           a string of random bytes of the given length. Defaults
      *                                                           to the `random_bytes()` function.
@@ -269,8 +269,8 @@ final readonly class BigInteger extends BigNumber
      *                       or `$min` is greater than `$max`.
      */
     public static function randomRange(
-        BigNumber|int|float|string $min,
-        BigNumber|int|float|string $max,
+        BigNumber|int|string $min,
+        BigNumber|int|string $max,
         ?callable $randomBytesGenerator = null,
     ): BigInteger {
         $min = BigInteger::of($min);
@@ -347,14 +347,14 @@ final readonly class BigInteger extends BigNumber
     }
 
     /**
-     * @param BigNumber|int|float|string $a    The first number. Must be convertible to a BigInteger.
-     * @param BigNumber|int|float|string ...$n The subsequent numbers. Must be convertible to BigInteger.
+     * @param BigNumber|int|string $a    The first number. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string ...$n The subsequent numbers. Must be convertible to BigInteger.
      *
      * @throws MathException If one of the parameters cannot be converted to a BigInteger.
      *
      * @pure
      */
-    public static function gcdAll(BigNumber|int|float|string $a, BigNumber|int|float|string ...$n): BigInteger
+    public static function gcdAll(BigNumber|int|string $a, BigNumber|int|string ...$n): BigInteger
     {
         $result = BigInteger::of($a);
 
@@ -372,13 +372,13 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the sum of this number and the given one.
      *
-     * @param BigNumber|int|float|string $that The number to add. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The number to add. Must be convertible to a BigInteger.
      *
      * @throws MathException If the number is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function plus(BigNumber|int|float|string $that): BigInteger
+    public function plus(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -398,13 +398,13 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the difference of this number and the given one.
      *
-     * @param BigNumber|int|float|string $that The number to subtract. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The number to subtract. Must be convertible to a BigInteger.
      *
      * @throws MathException If the number is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function minus(BigNumber|int|float|string $that): BigInteger
+    public function minus(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -420,13 +420,13 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the product of this number and the given one.
      *
-     * @param BigNumber|int|float|string $that The multiplier. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The multiplier. Must be convertible to a BigInteger.
      *
      * @throws MathException If the multiplier is not a valid number, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function multipliedBy(BigNumber|int|float|string $that): BigInteger
+    public function multipliedBy(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -446,15 +446,15 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the result of the division of this number by the given one.
      *
-     * @param BigNumber|int|float|string $that         The divisor. Must be convertible to a BigInteger.
-     * @param RoundingMode               $roundingMode An optional rounding mode, defaults to Unnecessary.
+     * @param BigNumber|int|string $that         The divisor. Must be convertible to a BigInteger.
+     * @param RoundingMode         $roundingMode An optional rounding mode, defaults to Unnecessary.
      *
      * @throws MathException If the divisor is not a valid number, is not convertible to a BigInteger, is zero,
      *                       or RoundingMode::Unnecessary is used and the remainder is not zero.
      *
      * @pure
      */
-    public function dividedBy(BigNumber|int|float|string $that, RoundingMode $roundingMode = RoundingMode::Unnecessary): BigInteger
+    public function dividedBy(BigNumber|int|string $that, RoundingMode $roundingMode = RoundingMode::Unnecessary): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -478,13 +478,13 @@ final readonly class BigInteger extends BigNumber
      * If the number is greater than $max, returns a copy of $max.
      * Otherwise, returns this number unchanged.
      *
-     * @param BigNumber|int|float|string $min The minimum. Must be convertible to a BigInteger.
-     * @param BigNumber|int|float|string $max The maximum. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $min The minimum. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $max The maximum. Must be convertible to a BigInteger.
      *
      * @throws MathException            If min/max are not convertible to a BigInteger.
      * @throws InvalidArgumentException If min is greater than max.
      */
-    public function clamp(BigNumber|int|float|string $min, BigNumber|int|float|string $max): BigInteger
+    public function clamp(BigNumber|int|string $min, BigNumber|int|string $max): BigInteger
     {
         $min = BigInteger::of($min);
         $max = BigInteger::of($max);
@@ -533,14 +533,14 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the quotient of the division of this number by the given one.
      *
-     * @param BigNumber|int|float|string $that The divisor. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The divisor. Must be convertible to a BigInteger.
      *
      * @throws MathException           If the divisor is not valid, or is not convertible to a BigInteger.
      * @throws DivisionByZeroException If the divisor is zero.
      *
      * @pure
      */
-    public function quotient(BigNumber|int|float|string $that): BigInteger
+    public function quotient(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -562,14 +562,14 @@ final readonly class BigInteger extends BigNumber
      *
      * The remainder, when non-zero, has the same sign as the dividend.
      *
-     * @param BigNumber|int|float|string $that The divisor. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The divisor. Must be convertible to a BigInteger.
      *
      * @throws MathException           If the divisor is not valid, or is not convertible to a BigInteger.
      * @throws DivisionByZeroException If the divisor is zero.
      *
      * @pure
      */
-    public function remainder(BigNumber|int|float|string $that): BigInteger
+    public function remainder(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -589,7 +589,7 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the quotient and remainder of the division of this number by the given one.
      *
-     * @param BigNumber|int|float|string $that The divisor. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The divisor. Must be convertible to a BigInteger.
      *
      * @return array{BigInteger, BigInteger} An array containing the quotient and the remainder.
      *
@@ -598,7 +598,7 @@ final readonly class BigInteger extends BigNumber
      *
      * @pure
      */
-    public function quotientAndRemainder(BigNumber|int|float|string $that): array
+    public function quotientAndRemainder(BigNumber|int|string $that): array
     {
         $that = BigInteger::of($that);
 
@@ -622,14 +622,14 @@ final readonly class BigInteger extends BigNumber
      *
      * The result of the modulo operation, when non-zero, has the same sign as the divisor.
      *
-     * @param BigNumber|int|float|string $that The divisor. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $that The divisor. Must be convertible to a BigInteger.
      *
      * @throws MathException           If the divisor is not valid, or is not convertible to a BigInteger.
      * @throws DivisionByZeroException If the divisor is zero.
      *
      * @pure
      */
-    public function mod(BigNumber|int|float|string $that): BigInteger
+    public function mod(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -645,7 +645,7 @@ final readonly class BigInteger extends BigNumber
     /**
      * Returns the modular multiplicative inverse of this BigInteger modulo $modulus.
      *
-     * @param BigNumber|int|float|string $modulus The modulus. Must be convertible to a BigInteger.
+     * @param BigNumber|int|string $modulus The modulus. Must be convertible to a BigInteger.
      *
      * @throws DivisionByZeroException If $modulus is zero.
      * @throws NegativeNumberException If $modulus is negative.
@@ -654,7 +654,7 @@ final readonly class BigInteger extends BigNumber
      *
      * @pure
      */
-    public function modInverse(BigNumber|int|float|string $modulus): BigInteger
+    public function modInverse(BigNumber|int|string $modulus): BigInteger
     {
         $modulus = BigInteger::of($modulus);
 
@@ -684,8 +684,8 @@ final readonly class BigInteger extends BigNumber
      *
      * This operation only works on positive numbers.
      *
-     * @param BigNumber|int|float|string $exponent The exponent. Must be positive or zero.
-     * @param BigNumber|int|float|string $modulus  The modulus. Must be strictly positive.
+     * @param BigNumber|int|string $exponent The exponent. Must be positive or zero.
+     * @param BigNumber|int|string $modulus  The modulus. Must be strictly positive.
      *
      * @throws MathException           If the exponent or modulus is not valid, or not convertible to a BigInteger.
      * @throws NegativeNumberException If any of the operands is negative.
@@ -693,7 +693,7 @@ final readonly class BigInteger extends BigNumber
      *
      * @pure
      */
-    public function modPow(BigNumber|int|float|string $exponent, BigNumber|int|float|string $modulus): BigInteger
+    public function modPow(BigNumber|int|string $exponent, BigNumber|int|string $modulus): BigInteger
     {
         $exponent = BigInteger::of($exponent);
         $modulus = BigInteger::of($modulus);
@@ -716,13 +716,13 @@ final readonly class BigInteger extends BigNumber
      *
      * The GCD is always positive, unless both operands are zero, in which case it is zero.
      *
-     * @param BigNumber|int|float|string $that The operand. Must be convertible to an integer number.
+     * @param BigNumber|int|string $that The operand. Must be convertible to an integer number.
      *
      * @throws MathException If the operand is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function gcd(BigNumber|int|float|string $that): BigInteger
+    public function gcd(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -828,13 +828,13 @@ final readonly class BigInteger extends BigNumber
      *
      * This method returns a negative BigInteger if and only if both operands are negative.
      *
-     * @param BigNumber|int|float|string $that The operand. Must be convertible to an integer number.
+     * @param BigNumber|int|string $that The operand. Must be convertible to an integer number.
      *
      * @throws MathException If the operand is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function and(BigNumber|int|float|string $that): BigInteger
+    public function and(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -846,13 +846,13 @@ final readonly class BigInteger extends BigNumber
      *
      * This method returns a negative BigInteger if and only if either of the operands is negative.
      *
-     * @param BigNumber|int|float|string $that The operand. Must be convertible to an integer number.
+     * @param BigNumber|int|string $that The operand. Must be convertible to an integer number.
      *
      * @throws MathException If the operand is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function or(BigNumber|int|float|string $that): BigInteger
+    public function or(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -864,13 +864,13 @@ final readonly class BigInteger extends BigNumber
      *
      * This method returns a negative BigInteger if and only if exactly one of the operands is negative.
      *
-     * @param BigNumber|int|float|string $that The operand. Must be convertible to an integer number.
+     * @param BigNumber|int|string $that The operand. Must be convertible to an integer number.
      *
      * @throws MathException If the operand is not valid, or is not convertible to a BigInteger.
      *
      * @pure
      */
-    public function xor(BigNumber|int|float|string $that): BigInteger
+    public function xor(BigNumber|int|string $that): BigInteger
     {
         $that = BigInteger::of($that);
 
@@ -1018,7 +1018,7 @@ final readonly class BigInteger extends BigNumber
     }
 
     #[Override]
-    public function compareTo(BigNumber|int|float|string $that): int
+    public function compareTo(BigNumber|int|string $that): int
     {
         $that = BigNumber::of($that);