+++ src/BigNumber.php
@@ -16,9 +16,7 @@ use Stringable;
 use function array_shift;
 use function assert;
 use function filter_var;
-use function is_float;
 use function is_int;
-use function is_nan;
 use function is_null;
 use function ltrim;
 use function preg_match;
@@ -26,9 +24,7 @@ use function str_contains;
 use function str_repeat;
 use function strlen;
 use function substr;
-use function trigger_error;
 
-use const E_USER_DEPRECATED;
 use const FILTER_VALIDATE_INT;
 use const PREG_UNMATCHED_AS_NULL;
 
@@ -73,7 +69,6 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * - BigNumber instances are returned as is
      * - integer numbers are returned as BigInteger
-     * - floating point numbers are converted to a string then parsed as such (deprecated, will be removed in 0.15)
      * - strings containing a `/` character are returned as BigRational
      * - strings containing a `.` character or using an exponential notation are returned as BigDecimal
      * - strings containing only digits with an optional leading `+` or `-` sign are returned as BigInteger
@@ -87,7 +82,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public static function of(BigNumber|int|float|string $value): static
+    final public static function of(BigNumber|int|string $value): static
     {
         $value = self::_of($value);
 
@@ -111,7 +106,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      * @throws DivisionByZeroException    If the value represents a rational number with a denominator of zero.
      * @throws RoundingNecessaryException If the value cannot be converted to an instance of the subclass without rounding.
      */
-    public static function ofNullable(BigNumber|int|float|string|null $value): ?static
+    public static function ofNullable(BigNumber|int|string|null $value): ?static
     {
         if (is_null($value)) {
             return null;
@@ -123,15 +118,15 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
     /**
      * Returns the minimum of the given values.
      *
-     * @param BigNumber|int|float|string ...$values The numbers to compare. All the numbers need to be convertible
-     *                                              to an instance of the class this method is called on.
+     * @param BigNumber|int|string ...$values The numbers to compare. All the numbers need to be convertible
+     *                                        to an instance of the class this method is called on.
      *
      * @throws InvalidArgumentException If no values are given.
      * @throws MathException            If an argument is not valid.
      *
      * @pure
      */
-    final public static function min(BigNumber|int|float|string ...$values): static
+    final public static function min(BigNumber|int|string ...$values): static
     {
         $min = null;
 
@@ -153,15 +148,15 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
     /**
      * Returns the maximum of the given values.
      *
-     * @param BigNumber|int|float|string ...$values The numbers to compare. All the numbers need to be convertible
-     *                                              to an instance of the class this method is called on.
+     * @param BigNumber|int|string ...$values The numbers to compare. All the numbers need to be convertible
+     *                                        to an instance of the class this method is called on.
      *
      * @throws InvalidArgumentException If no values are given.
      * @throws MathException            If an argument is not valid.
      *
      * @pure
      */
-    final public static function max(BigNumber|int|float|string ...$values): static
+    final public static function max(BigNumber|int|string ...$values): static
     {
         $max = null;
 
@@ -189,15 +184,15 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      * When called on BigInteger, BigDecimal, or BigRational, sum() requires that all values can be converted to that
      * specific subclass, and returns a result of the same type.
      *
-     * @param BigNumber|int|float|string ...$values The values to add. All values must be convertible to the class on
-     *                                              which this method is called.
+     * @param BigNumber|int|string ...$values The values to add. All values must be convertible to the class on
+     *                                        which this method is called.
      *
      * @throws InvalidArgumentException If no values are given.
      * @throws MathException            If an argument is not valid.
      *
      * @pure
      */
-    final public static function sum(BigNumber|int|float|string ...$values): static
+    final public static function sum(BigNumber|int|string ...$values): static
     {
         $first = array_shift($values);
 
@@ -223,7 +218,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public function isEqualTo(BigNumber|int|float|string $that): bool
+    final public function isEqualTo(BigNumber|int|string $that): bool
     {
         return $this->compareTo($that) === 0;
     }
@@ -235,7 +230,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public function isLessThan(BigNumber|int|float|string $that): bool
+    final public function isLessThan(BigNumber|int|string $that): bool
     {
         return $this->compareTo($that) < 0;
     }
@@ -247,7 +242,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public function isLessThanOrEqualTo(BigNumber|int|float|string $that): bool
+    final public function isLessThanOrEqualTo(BigNumber|int|string $that): bool
     {
         return $this->compareTo($that) <= 0;
     }
@@ -259,7 +254,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public function isGreaterThan(BigNumber|int|float|string $that): bool
+    final public function isGreaterThan(BigNumber|int|string $that): bool
     {
         return $this->compareTo($that) > 0;
     }
@@ -271,7 +266,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    final public function isGreaterThanOrEqualTo(BigNumber|int|float|string $that): bool
+    final public function isGreaterThanOrEqualTo(BigNumber|int|string $that): bool
     {
         return $this->compareTo($that) >= 0;
     }
@@ -348,7 +343,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    abstract public function compareTo(BigNumber|int|float|string $that): int;
+    abstract public function compareTo(BigNumber|int|string $that): int;
 
     /**
      * Converts this number to a BigInteger.
@@ -489,7 +484,7 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
      *
      * @pure
      */
-    private static function _of(BigNumber|int|float|string $value): BigNumber
+    private static function _of(BigNumber|int|string $value): BigNumber
     {
         if ($value instanceof BigNumber) {
             return $value;
@@ -497,21 +492,6 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
 
         if (is_int($value)) {
             return new BigInteger((string) $value);
-        }
-
-        if (is_float($value)) {
-            // @phpstan-ignore-next-line
-            trigger_error(
-                'Passing floats to BigNumber::of() and arithmetic methods is deprecated and will be removed in 0.15. ' .
-                'Cast the float to string explicitly to preserve the previous behaviour.',
-                E_USER_DEPRECATED,
-            );
-
-            if (is_nan($value)) {
-                $value = 'NAN';
-            } else {
-                $value = (string) $value;
-            }
         }
 
         if (str_contains($value, '/')) {